Customized tests to support Rubahs testing framework

+++ h2/src/test/org/h2/test/db/TestRights.java
@@ -17,6 +17,8 @@ import org.h2.constant.ErrorCode;
 import org.h2.test.TestBase;
 import org.h2.util.FileUtils;
 
+import rubah.test.Test;
+
 /**
  * Access rights tests.
  */
@@ -37,7 +39,7 @@ public class TestRights extends TestBase {
         testDropOwnUser();
         testGetTables();
         testDropTempTables();
-        // testLowerCaseUser();
+//        // testLowerCaseUser();
         testSchemaRenameUser();
         testAccessRights();
         deleteDb("rights");
@@ -54,17 +56,41 @@ public class TestRights extends TestBase {
         } catch (SQLException e) {
             assertEquals(ErrorCode.CANNOT_DROP_CURRENT_USER, e.getErrorCode());
         }
+
+
+        Test.allowUpdates();
         stat.execute("CREATE USER TEST PASSWORD 'TEST' ADMIN");
+        Connection keepAlive = null;
+
+        // Keep a connection so memory database is not deleted
+        if (config.memory)
+        	keepAlive = getConnection("rights", "TEST", "TEST");
+
+
+        if (!Test.updated()) {
+        	// Update has not happened yet and rest of test requires the new version
+        	// Quit this execution
+        	Test.disallowUpdates();
+        	conn.close();
+        	if (config.memory)
+        		keepAlive.close();
+        	return;
+        }
+
         stat.execute("DROP USER " + user);
         conn.close();
-        if (!config.memory) {
-            try {
-                getConnection("rights");
-                fail();
-            } catch (SQLException e) {
-                assertKnownException(e);
-            }
+
+        try {
+        	getConnection("rights");
+        	fail();
+        } catch (SQLException e) {
+        	assertKnownException(e);
         }
+
+        Test.disallowUpdates();
+
+        if (config.memory)
+        	keepAlive.close();
     }
 
 //    public void testLowerCaseUser() throws SQLException {
@@ -87,12 +113,14 @@ public class TestRights extends TestBase {
         Connection conn = getConnection("rights");
         stat = conn.createStatement();
 
+        Test.allowUpdates();
         stat.execute("CREATE USER IF NOT EXISTS TEST PASSWORD 'TEST'");
         stat.execute("CREATE TABLE TEST(ID INT)");
         stat.execute("GRANT ALL ON TEST TO TEST");
         Connection conn2 = getConnection("rights", "TEST", getPassword("TEST"));
         DatabaseMetaData meta = conn2.getMetaData();
         meta.getTables(null, null, "%", new String[]{"TABLE", "VIEW", "SEQUENCE"});
+        Test.disallowUpdates();
         conn2.close();
         conn.close();
     }
@@ -105,6 +133,7 @@ public class TestRights extends TestBase {
         stat.execute("CREATE TABLE TEST(ID INT)");
         Connection conn2 = getConnection("rights", "READER", getPassword("READER"));
         Statement stat2 = conn2.createStatement();
+        Test.allowUpdates();
         try {
             stat2.execute("SELECT * FROM TEST");
             fail();
@@ -115,6 +144,7 @@ public class TestRights extends TestBase {
         stat2.execute("INSERT INTO MY_TEST VALUES(1)");
         stat2.execute("SELECT * FROM MY_TEST");
         stat2.execute("DROP TABLE MY_TEST");
+        Test.disallowUpdates();
         conn2.close();
         conn.close();
     }
